[VSC-6] Add benchmark for hover
